# openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [
    { "url": "http://localhost:8086", "description": "Generated server url" }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/payment/{bookingId}": {
      "post": {
        "tags": ["payment-cantroller"],
        "summary": "Create Booking",
        "operationId": "randomPaymentStatus",
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "201": {
            "description": "Booking Created",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/PaymentResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "status": { "type": "integer", "format": "int32" },
          "errorCode": { "type": "string" },
          "message": { "type": "string" },
          "path": { "type": "string" },
          "traceId": { "type": "string" }
        }
      },
      "PaymentResponse": {
        "type": "object",
        "properties": {
          "paymentId": { "type": "string", "format": "uuid" },
          "bookingId": { "type": "string", "format": "uuid" },
          "amount": { "type": "number" },
          "status": {
            "type": "string",
            "enum": ["PENDING", "SUCCESS", "FAILED", "REFUNDED", "EXPIRED"]
          },
          "gatewayPaymentId": { "type": "string" },
          "failureReason": { "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}